applied new timer - without window patching
