+++ js/timer/time-format.js
@@ -1,4 +1,4 @@
-window.formatTime = (total, passed) => {
+export default (total, passed) => {
   const minutesLeft = Math.floor((total - passed) / 60 / 1000);
   const secondsLeft = (total - passed - minutesLeft * 60 * 1000) / 1000;
